some changes to files

+++ lib/core/database/tables/transactions_table.dart
@@ -8,7 +8,7 @@ class TransactionsTable extends Table {
   TextColumn get fromUserPublicKey =>
       text().references(UsersTable, #publicKey)();
   TextColumn get toUserPublicKey => text().references(UsersTable, #publicKey)();
-  IntColumn get amount => integer()();
+  Int64Column get amount => int64()();
   TextColumn get description => text().nullable()();
   TextColumn get tag => text().nullable()();
   TextColumn get status => textEnum<TransactionStatus>()();
